+++ packages/core/src/sessions/http/utils/get-csrf-token-from-request.ts
@@ -1,4 +1,4 @@
-import { Context } from '../../core';
+import { Context } from '../../../core';
 
 export function getCsrfTokenFromRequest(request: Context['request']): string|undefined {
   return request.body._csrf || request.get('X-CSRF-Token') || request.get('X-XSRF-Token');
