+++ packages/core/src/sessions/http/utils/index.ts
@@ -1,1 +1,4 @@
+export { getCsrfTokenFromRequest } from './get-csrf-token-from-request';
 export { shouldVerifyCsrfToken } from './should-verify-csrf-token';
+export { removeSessionCookie } from './remove-session-cookie';
+export { setSessionCookie } from './set-session-cookie';
