+++ packages/jwt/src/http/jwt.hook.ts
@@ -20,9 +20,8 @@ import { decode, verify } from 'jsonwebtoken';
 // FoalTS
 import { JWT_DEFAULT_COOKIE_NAME, JWT_DEFAULT_CSRF_COOKIE_NAME } from './constants';
 import { getSecretOrPublicKey } from '../core';
-import { getCsrfTokenFromRequest } from './get-csrf-token-from-request';
 import { isInvalidTokenError } from './invalid-token.error';
-import { getCsrfTokenFromCookie, shouldVerifyCsrfToken } from './utils';
+import { getCsrfTokenFromCookie, getCsrfTokenFromRequest, shouldVerifyCsrfToken } from './utils';
 
 class InvalidTokenResponse extends HttpResponseUnauthorized {
