+++ packages/jwt/src/http/utils/get-csrf-token-from-request.ts
