+++ packages/jwt/src/http/utils/index.ts
@@ -1,2 +1,3 @@
 export { getCsrfTokenFromCookie } from './get-csrf-token-from-cookie';
+export { getCsrfTokenFromRequest } from './get-csrf-token-from-request';
 export { shouldVerifyCsrfToken } from './should-verify-csrf-token';
